[merged] with private branch
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3devel/MMOCoreORB@157 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/objects/player/Player.h
@@ -377,7 +377,7 @@ public:
 
 	Datapad* getDatapad();
 
-	unsigned int getNewItemID();
+	unsigned long long getNewItemID();
 
 	unsigned int getItemShift();
 
@@ -732,7 +732,7 @@ public:
 
 	Datapad* getDatapad();
 
-	unsigned int getNewItemID();
+	unsigned long long getNewItemID();
 
 	unsigned int getItemShift();
